+++ lib/generators/templates/rapns.rb
@@ -63,6 +63,11 @@ Rapns.reflect do |on|
   # on.apns_connection_lost do |app, error|
   # end
 
+  # Called when the GCM returns a canonical registration ID.
+  # You will need to replace old_id with canonical_id in your records.
+  # on.gcm_canonical_id do |old_id, canonical_id|
+  # end
+
   # Called when an exception is raised.
   # on.error do |error|
   # end