+++ lib/rapns/daemon.rb
@@ -8,7 +8,6 @@ require 'net/http/persistent'
 require 'rapns/daemon/reflectable'
 require 'rapns/daemon/interruptible_sleep'
 require 'rapns/daemon/delivery_error'
-require 'rapns/daemon/database_reconnectable'
 require 'rapns/daemon/delivery'
 require 'rapns/daemon/delivery_queue'
 require 'rapns/daemon/feeder'
@@ -28,14 +27,19 @@ require 'rapns/daemon/gcm/delivery_handler'
 
 module Rapns
   module Daemon
-    extend DatabaseReconnectable
+    class << self
+      attr_accessor :store
+    end
 
     def self.start
       setup_signal_traps if trap_signals?
 
+      initialize_store
+      return unless store
+
       if daemonize?
         daemonize
-        reconnect_database
+        store.after_daemonize
       end
 
       write_pid_file
@@ -51,10 +55,22 @@ module Rapns
       delete_pid_file
     end
 
+    def self.initialize_store
+      return if store
+      begin
+        require "rapns/daemon/store/#{Rapns.config.store}"
+        klass = "Rapns::Daemon::Store::#{Rapns.config.store.to_s.camelcase}".constantize
+        self.store = klass.new
+      rescue StandardError, LoadError => e
+        Rapns.logger.error("Failed to load '#{Rapns.config.store}' storage backend.")
+        Rapns.logger.error(e)
+      end
+    end
+
     protected
 
     def self.daemonize?
-      !(Rapns.config.foreground || Rapns.config.embedded || defined?(JRUBY_VERSION))
+      !(Rapns.config.foreground || Rapns.config.embedded || Rapns.jruby?)
     end
 
     def self.trap_signals?