+++ lib/rapns/daemon/delivery.rb
@@ -1,7 +1,6 @@
 module Rapns
   module Daemon
     class Delivery
-      include DatabaseReconnectable
       include Reflectable
 
       def self.perform(*args)
@@ -9,11 +8,7 @@ module Rapns
       end
 
       def retry_after(notification, deliver_after)
-        with_database_reconnect_and_retry do
-          notification.retries += 1
-          notification.deliver_after = deliver_after
-          notification.save!(:validate => false)
-        end
+        Rapns::Daemon.store.retry_after(notification, deliver_after)
         reflect(:notification_will_retry, notification)
       end
 
@@ -22,24 +17,12 @@ module Rapns
       end
 
       def mark_delivered
-        with_database_reconnect_and_retry do
-          @notification.delivered = true
-          @notification.delivered_at = Time.now
-          @notification.save!(:validate => false)
-        end
+        Rapns::Daemon.store.mark_delivered(@notification)
         reflect(:notification_delivered, @notification)
       end
 
       def mark_failed(code, description)
-        with_database_reconnect_and_retry do
-          @notification.delivered = false
-          @notification.delivered_at = nil
-          @notification.failed = true
-          @notification.failed_at = Time.now
-          @notification.error_code = code
-          @notification.error_description = description
-          @notification.save!(:validate => false)
-        end
+        Rapns::Daemon.store.mark_failed(@notification, code, description)
         reflect(:notification_failed, @notification)
       end
     end