+++ spec/unit/daemon/feeder_spec.rb
@@ -8,10 +8,10 @@ describe Rapns::Daemon::Feeder do
   let(:logger) { stub }
 
   before do
-    Rapns.stub(:config => config)
+    Rapns.stub(:config => config,:logger => logger)
+    Rapns::Daemon.stub(:store => stub(:deliverable_notifications => [notification]))
     Rapns::Daemon::Feeder.stub(:stop? => true)
-    Rapns::Daemon::AppRunner.stub(:idle => [stub(:app => app)])
-    Rapns.stub(:logger => logger)
+    Rapns::Daemon::AppRunner.stub(:enqueue => nil, :idle => [stub(:app => app)])
   end
 
   def start
@@ -25,8 +25,8 @@ describe Rapns::Daemon::Feeder do
     start
   end
 
-  it "checks for new notifications with the ability to reconnect the database" do
-    Rapns::Daemon::Feeder.should_receive(:with_database_reconnect_and_retry)
+  it 'loads deliverable notifications' do
+    Rapns::Daemon.store.should_receive(:deliverable_notifications).with([app])
     start
   end
 
@@ -37,72 +37,20 @@ describe Rapns::Daemon::Feeder do
     start
   end
 
-  it 'loads notifications in batches' do
-    relation = stub.as_null_object
-    relation.should_receive(:limit).with(5000)
-    Rapns::Notification.stub(:ready_for_delivery => relation)
-    start
-  end
-
-  it 'does not load notification in batches if in push mode' do
-    config.stub(:push => true)
-    relation = stub.as_null_object
-    relation.should_not_receive(:limit)
-    Rapns::Notification.stub(:ready_for_delivery => relation)
-    start
-  end
-
   it "enqueues the notification" do
-    notification.update_attributes!(:delivered => false)
     Rapns::Daemon::AppRunner.should_receive(:enqueue).with(notification)
     start
   end
 
   it 'reflects the notification has been enqueued' do
-    notification.update_attributes!(:delivered => false)
     Rapns::Daemon::AppRunner.stub(:enqueue)
     Rapns::Daemon::Feeder.should_receive(:reflect).with(:notification_enqueued, notification)
     start
   end
 
-  it 'does not enqueue the notification if the app runner is still processing the previous batch' do
-    Rapns::Daemon::AppRunner.should_not_receive(:enqueue)
-    start
-  end
-
-  it "enqueues an undelivered notification without deliver_after set" do
-    notification.update_attributes!(:delivered => false, :deliver_after => nil)
-    Rapns::Daemon::AppRunner.should_receive(:enqueue).with(notification)
-    start
-  end
-
-  it "enqueues a notification with a deliver_after time in the past" do
-    notification.update_attributes!(:delivered => false, :deliver_after => 1.hour.ago)
-    Rapns::Daemon::AppRunner.should_receive(:enqueue).with(notification)
-    start
-  end
-
-  it "does not enqueue a notification with a deliver_after time in the future" do
-    notification.update_attributes!(:delivered => false, :deliver_after => 1.hour.from_now)
-    Rapns::Daemon::AppRunner.should_not_receive(:enqueue)
-    start
-  end
-
-  it "does not enqueue a previously delivered notification" do
-    notification.update_attributes!(:delivered => true, :delivered_at => Time.now)
-    Rapns::Daemon::AppRunner.should_not_receive(:enqueue)
-    start
-  end
-
-  it "does not enqueue a notification that has previously failed delivery" do
-    notification.update_attributes!(:delivered => false, :failed => true)
-    Rapns::Daemon::AppRunner.should_not_receive(:enqueue)
-    start
-  end
-
   it "logs errors" do
     e = StandardError.new("bork")
-    Rapns::Notification.stub(:ready_for_delivery).and_raise(e)
+    Rapns::Daemon.store.stub(:deliverable_notifications).and_raise(e)
     Rapns.logger.should_receive(:error).with(e)
     start
   end