+++ spec/unit/daemon_spec.rb
@@ -1,4 +1,5 @@
 require 'unit_spec_helper'
+require 'rapns/daemon/store/active_record'
 
 describe Rapns::Daemon, "when starting" do
   module Rails; end
@@ -6,24 +7,31 @@ describe Rapns::Daemon, "when starting" do
   let(:certificate) { stub }
   let(:password) { stub }
   let(:config) { stub(:pid_file => nil, :airbrake_notify => false,
-    :foreground => true, :embedded => false, :push => false) }
+    :foreground => true, :embedded => false, :push => false, :store => :active_record) }
   let(:logger) { stub(:logger, :info => nil, :error => nil, :warn => nil) }
 
   before do
-    Rapns.stub(:config => config)
-    Rapns::Logger.stub(:new => logger)
+    Rapns.stub(:config => config, :logger => logger)
     Rapns::Daemon::Feeder.stub(:start)
     Rapns::Daemon::AppRunner.stub(:sync => nil, :stop => nil)
-    Rapns::Daemon.stub(:daemonize => nil, :reconnect_database => nil, :exit => nil, :puts => nil)
+    Rapns::Daemon.stub(:daemonize => nil, :exit => nil, :puts => nil)
     File.stub(:open)
     Rails.stub(:root).and_return("/rails_root")
   end
 
-  unless defined?(JRUBY_VERSION)
+  unless Rapns.jruby?
     it "forks into a daemon if the foreground option is false" do
       config.stub(:foreground => false)
-      ActiveRecord::Base.stub(:establish_connection)
+      Rapns::Daemon.initialize_store
+      Rapns::Daemon.store.stub(:after_daemonize => nil)
       Rapns::Daemon.should_receive(:daemonize)
+      Rapns::Daemon.start
+    end
+
+    it 'notifies the store after forking' do
+      config.stub(:foreground => false)
+      Rapns::Daemon.initialize_store
+      Rapns::Daemon.store.should_receive(:after_daemonize)
       Rapns::Daemon.start
     end
 
@@ -54,6 +62,18 @@ describe Rapns::Daemon, "when starting" do
   it 'does not setup signal traps when embedded' do
     config.stub(:embedded => true)
     Rapns::Daemon.should_not_receive(:setup_signal_traps)
+    Rapns::Daemon.start
+  end
+
+  it 'instantiates the store' do
+    config.stub(:store => :active_record)
+    Rapns::Daemon.start
+    Rapns::Daemon.store.should be_kind_of(Rapns::Daemon::Store::ActiveRecord)
+  end
+
+  it 'logs an error if the store cannot be loaded' do
+    config.stub(:store => :foo_bar)
+    Rapns.logger.should_receive(:error).with(kind_of(LoadError))
     Rapns::Daemon.start
   end
 
@@ -91,7 +111,7 @@ describe Rapns::Daemon, "when being shutdown" do
   end
 
   # These tests do not work on JRuby.
-  unless defined? JRUBY_VERSION
+  unless Rapns.jruby?
     it "shuts down when signaled signaled SIGINT" do
       Rapns::Daemon.setup_signal_traps
       Rapns::Daemon.should_receive(:shutdown)